chore: update testdata
This updates the `urlpatterntestdata.json` file to match the WPT.

One of the test cases added in web-platform-tests/wpt#30484 has
`.expected_match.input` be an array of one element, which by default
won't deserialize a 2-element tuple. This change fixes that by providing
a custom deserialization function. Other test cases had `pattern` be an
empty array, which caused panics in the `test_case()` function. This
change also fixes that.